move to domain based hook for extension and start breaking hook apart
