--- src/hooks/use-extension.js
@@ -1,273 +1,0 @@
-import { useEffect, useState } from 'react'
-import {
-  ALARM_KEY,
-  TOP_SITES_KEY,
-  TOP_SITES_USAGE_KEY,
-  PAGE_VIEWS_KEY,
-  SITE_TIME_KEY,
-  SHOW_TOP_SITES_KEY,
-  NEW_TAB_CONNECTION,
-  BREAK_TIMERS_KEY,
-  RESET_USAGE,
-  ACTIVITY_NUMBER_KEY,
-  ADD_BROADCAST_TAB,
-  VALUE_CHANGED,
-  MAX_BREAKTIMER_MODAL,
-  MAX_BREAKTIMERS,
-  SUB_DETAILS_MODAL,
-  CURRENT_SUB_SKU,
-  APP_READY_KEY,
-  SCREEN_TIME_PERMISSIONS,
-  SUPPORT_URL,
-} from '../constants'
-import { toSiteInfo } from '../lib/aggregation'
-import {
-  message,
-  topSites as topSitesApi,
-  onMessage,
-} from '../lib/extension'
-import { first } from '../lib/functional'
-import { queryParams, shortURL, origin } from '../lib/url'
-import { set, create } from '../lib/util'
-import { usePermissions } from './use-permissions'
-import { useStorage } from './use-storage'
-import { useSubscription } from './use-subscription'
-
-const REMINDER_ALT = 'Notifications that remind you to take a break'
-
-export const onBackgroundMessage = updateFns => payload => {
-  const { key, event } = payload
-  if (event !== VALUE_CHANGED) return
-  const fn = updateFns[key]
-  if (typeof fn === 'function') {
-    fn(null, { refresh: true })
-  }
-}
-
-export const useExtension = () => {
-  const [appReady, setAppReady] = useState(false)
-  const [playerIsOpen, setPlayerIsOpen] = useState(false)
-  const [selectedSegment, setSelectedSegment] = useState(null)
-  const [upsellModal, setUpsellModal] = useState(null)
-  const [alarmEnabled, setAlarmEnabled] = useStorage(ALARM_KEY)
-  const [topSites, setTopSites] = useStorage(TOP_SITES_KEY)
-  const [siteTimes, setSiteTimes] = useStorage(SITE_TIME_KEY)
-  const [topSitesUsage, setTopSitesUsage] = useStorage(
-    TOP_SITES_USAGE_KEY
-  )
-  const [pageViews, updatePageViews, { pending }] = useStorage(
-    PAGE_VIEWS_KEY
-  )
-  const [showTopSites, updateShowTopSites] = useStorage(
-    SHOW_TOP_SITES_KEY
-  )
-  const [breakTimers, updateBreakTimers] = useStorage(
-    BREAK_TIMERS_KEY
-  )
-  const [activityNumber, updateActivityNumber] = useStorage(
-    ACTIVITY_NUMBER_KEY
-  )
-  const { user } = useSubscription()
-  const {
-    hasPermission,
-    requestPermissions,
-    removePermissions,
-  } = usePermissions(SCREEN_TIME_PERMISSIONS)
-
-  const updateFns = {
-    [ALARM_KEY]: setAlarmEnabled,
-    [TOP_SITES_KEY]: setTopSites,
-    [SITE_TIME_KEY]: setSiteTimes,
-    [TOP_SITES_USAGE_KEY]: setTopSitesUsage,
-    [PAGE_VIEWS_KEY]: updatePageViews,
-    [SHOW_TOP_SITES_KEY]: updateShowTopSites,
-    [BREAK_TIMERS_KEY]: updateBreakTimers,
-    [ACTIVITY_NUMBER_KEY]: updateActivityNumber,
-  }
-
-  // mount hook
-  useEffect(() => {
-    if (!pending && !appReady) {
-      updatePageViews(pageViews + 1)
-      topSitesApi.get(setTopSites)
-      onMessage(onBackgroundMessage(updateFns))
-      message(NEW_TAB_CONNECTION)
-      message(ADD_BROADCAST_TAB)
-      setAppReady(true)
-      if (typeof window[APP_READY_KEY] === 'function') {
-        window[APP_READY_KEY]()
-      }
-      window[APP_READY_KEY] = true
-      const { site, play } = queryParams(window.location.href)
-      if (play) {
-        setPlayerIsOpen(true)
-      } else if (site) {
-        setSelectedSegment({
-          label: shortURL(site),
-          urls: [origin(site)],
-          link: site,
-        })
-      }
-    }
-  }, [
-    pageViews,
-    pending,
-    setSiteTimes,
-    setTopSites,
-    updatePageViews,
-    appReady,
-    updateFns,
-  ])
-
-  const setAlarmEnabledProxy = () => {
-    const enabled = !alarmEnabled
-    setAlarmEnabled(enabled)
-  }
-
-  const setBreakTimer = (url, time, enabled) => {
-    const lastBreakTimer = breakTimers[url] || {}
-    const nextBreakTimer = Object.assign({}, lastBreakTimer, {
-      url,
-      time,
-      enabled,
-    })
-    const nextBreakTimers = create(breakTimers, url, nextBreakTimer)
-    const enabledTimers = Object.keys(nextBreakTimers).filter(
-      key => nextBreakTimers[key].enabled
-    )
-    const overLimit = enabledTimers.length > MAX_BREAKTIMERS
-    if (overLimit && !user.isSubscribed) {
-      setUpsellModal({
-        name: MAX_BREAKTIMER_MODAL,
-        url: shortURL(url),
-        onClick: () =>
-          setUpsellModal({
-            name: SUB_DETAILS_MODAL,
-            sku: CURRENT_SUB_SKU,
-            callback: () => setBreakTimer(url, time, enabled),
-          }),
-      })
-      return
-    }
-    updateBreakTimers(nextBreakTimers)
-  }
-
-  const updateSitesUsed = site => {
-    setTopSitesUsage([...topSitesUsage, site])
-  }
-
-  const resetUsage = () => {
-    message(RESET_USAGE)
-  }
-
-  const mappedTopSites = topSites.map(topSite => {
-    const { url } = topSite
-    const isUsed = topSitesUsage.some(site => site.url === url)
-    return Object.assign({ isUsed }, topSite)
-  })
-
-  const siteTimesAndTimers = toSiteInfo(siteTimes, breakTimers)
-
-  const settings = [
-    {
-      label: 'Subscriber',
-      type: 'button',
-      value: 'More Info',
-      alt: user.isSubscribed
-        ? 'Thanks you for your support ❤️!'
-        : 'Get access to more Mujō',
-      setter: () => {
-        if (!user.isSubscribed) {
-          setUpsellModal({
-            name: SUB_DETAILS_MODAL,
-            sku: CURRENT_SUB_SKU,
-            callback: () => {},
-          })
-        } else {
-          setUpsellModal({
-            title: 'Contact Support',
-            description:
-              'We currently do not have this functionality',
-          })
-        }
-      },
-    },
-    {
-      label: 'Reminder',
-      type: 'boolean',
-      alt: REMINDER_ALT,
-      setter: enabled => setAlarmEnabled(enabled),
-      value: alarmEnabled,
-    },
-    {
-      label: 'Screen Time Enabled',
-      type: 'boolean',
-      alt: 'Screen Time requires some additional permissions',
-      setter: () => {
-        if (hasPermission) {
-          removePermissions()
-        } else {
-          requestPermissions()
-        }
-      },
-      value: hasPermission,
-    },
-    {
-      label: 'Help',
-      alt: 'Need help or got feedback? Talk to us on Spectrum.chat',
-      type: 'button',
-      value: 'Get support',
-      setter: () => {
-        window.location.href = SUPPORT_URL
-      },
-    },
-  ]
-
-  if (selectedSegment) {
-    const { urls } = selectedSegment
-    if (urls.length === 1) {
-      const originalURL = first(selectedSegment.urls)
-      const { breakTimer, time } = siteTimesAndTimers[originalURL]
-      set(selectedSegment, 'data', {
-        breakTimer,
-        time,
-        originalURL,
-      })
-    }
-  }
-
-  return [
-    {
-      topSites: mappedTopSites,
-      alarmEnabled,
-      pageViews,
-      showTopSites,
-      siteTimes,
-      siteTimesAndTimers,
-      appReady,
-      breakTimers,
-      selectedSegment,
-      playerIsOpen,
-      activityNumber,
-      upsellModal,
-      settings,
-      screenTime: {
-        hasPermission,
-        requestPermissions,
-        removePermissions,
-      },
-    },
-    {
-      setAlarmEnabled: setAlarmEnabledProxy,
-      setTopSites,
-      updateSitesUsed,
-      resetUsage,
-      updateShowTopSites,
-      setBreakTimer,
-      setSelectedSegment,
-      setPlayerIsOpen,
-      setUpsellModal,
-      updateBreakTimers,
-    },
-  ]
-}
