+++ src/index.js
@@ -4,6 +4,7 @@ import { useColorScheme } from 'use-color-scheme'
 import App from './app'
 import { ErrorBox } from './components/error-box'
 import { Font } from './components/fonts'
+import { ExtensionProvider } from './hooks/use-extension'
 import { SubscriptionProvider } from './hooks/use-subscription'
 import { ColorThemeProvider } from './hooks/use-theme'
 
@@ -15,9 +16,11 @@ const NewHomePage = () => {
     <ErrorBox>
       <Font />
       <SubscriptionProvider>
-        <ColorThemeProvider value={scheme}>
-          <App />
-        </ColorThemeProvider>
+        <ExtensionProvider>
+          <ColorThemeProvider value={scheme}>
+            <App />
+          </ColorThemeProvider>
+        </ExtensionProvider>
       </SubscriptionProvider>
     </ErrorBox>
   )
